aprimoramento

+++ Mundo-1/EXERCÍCIOS/EX014.py
@@ -1,5 +1,5 @@
 # Conversor de temperatura
 
-graC = int(input('Qual é a temperatura em °C: '))
+graC = float(input('Qual é a temperatura em °C: '))
 graF = graC * 1.8 + 32
 print('A temperatura °C:{} Convetida ficar °F:{}'.format(graC, graF))
